requirements.txt
asgiref==3.3.1
autopep8==1.5.4
beautifulsoup4==4.9.3
Django==3.1.5
django-bootstrap4==2.3.1
Pillow==8.1.0
pycodestyle==2.6.0
pytz==2020.5
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
